+++ code/part03/language/src/main/java/toyl/ast/ToylNode.java
@@ -4,5 +4,5 @@ import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.Node;
 
 public abstract class ToylNode extends Node {
-  public abstract long execute(VirtualFrame frame);
+  public abstract double execute(VirtualFrame frame);
 }
